requirements.txt
openai==1.58.1
openpyxl==3.1.5
pdf2image==1.17.0
python-dotenv==1.0.1
tenacity==9.0.0
pillow==10.4.0